Fix PR reviewer comments
